got abstraction api working
