+++ Fastload-1.19.0-1-2-Extension/src/main/resources/fastload-119-0-1-2-compat.mixins.json
@@ -5,7 +5,6 @@
   "plugin": "io.github.bumblesoftware.fastload.mixin.plugin.FLConditionalMixinConfig",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
-    "mixins.mc119.local.HookClient"
   ],
   "client": [
     "mixins.mc119.client.MinecraftClientMixin"
