+++ Fastload-1.19.3-Extension/src/main/java/io/github/bumblesoftware/fastload/abstraction/client/mc1193/Client1193.java
@@ -1,66 +1,36 @@
-package io.github.bumblesoftware.fastload.abstraction.client;
+package io.github.bumblesoftware.fastload.abstraction.client.mc1193;
+
 
 import com.mojang.serialization.Codec;
-import io.github.bumblesoftware.fastload.api.internal.abstraction.Client1182;
-import io.github.bumblesoftware.fastload.api.external.abstraction.tool.config.RetrieveValueFunction;
-import io.github.bumblesoftware.fastload.api.external.abstraction.tool.config.StoreValueFunction;
-import io.github.bumblesoftware.fastload.config.DefaultConfig;
-import io.github.bumblesoftware.fastload.util.Action;
+import io.github.bumblesoftware.fastload.abstraction.client.mc119.Client119;
+import io.github.bumblesoftware.fastload.api.external.abstraction.core.config.RetrieveValueFunction;
+import io.github.bumblesoftware.fastload.api.external.abstraction.core.config.StoreValueFunction;
 import io.github.bumblesoftware.fastload.util.Bound;
-import net.minecraft.client.gui.screen.Screen;
-import net.minecraft.client.gui.screen.option.SimpleOptionsScreen;
+import net.minecraft.client.gui.widget.ButtonWidget;
 import net.minecraft.client.option.GameOptions;
 import net.minecraft.client.option.SimpleOption;
 import net.minecraft.text.Text;
 
-import java.util.function.Function;
-
-public class Client119 extends Client1182 {
+public class Client1193 extends Client119 {
     @Override
     public String[] getSupportedVersions() {
-        return new String[] {"1.19", "1.19.1", "1.19.2"};
+        return new String[] {"1.19.3"};
     }
 
     @Override
-    public int getViewDistance() {
-        if (getClientInstance().options == null) {
-            return DefaultConfig.LOCAL_CHUNK_RADIUS_BOUND.max();
-        } else return getClientInstance().options.getClampedViewDistance();
-    }
-
-    @Override
-    public <T> Screen newConfigScreen(
-            final Screen parent,
-            GameOptions gameOptions,
-            final Text title,
-            Function<Object[], T[]> options,
-            Action config
+    public ButtonWidget getNewButton(
+            final int x,
+            final int y,
+            final int width,
+            final int height,
+            final Text message,
+            final ButtonWidget.PressAction onPress
     ) {
-        final var main = this;
-        return new SimpleOptionsScreen(
-                parent,
-                gameOptions,
-                title,
-                (SimpleOption<?>[]) options.apply(new SimpleOption[]{})
-        ) {
-            @Override
-            protected void initFooter() {
-                main.initFooter(this, parent, config);
-            }
-        };
+        return ButtonWidget.builder(message, onPress)
+                .dimensions(x, y, width, height)
+                .build();
     }
 
-    @Override
-    public Text newTranslatableText(final String content) {
-        return Text.translatable(content);
-    }
-
-    @Override
-    public Text newLiteralText(final String content) {
-        return Text.literal(content);
-    }
-
-    @SuppressWarnings("unchecked")
     @Override
     public SimpleOption<Boolean> newCyclingButton(
             final String namespace,
@@ -76,7 +46,6 @@ public class Client119 extends Client1182 {
         );
     }
 
-    @SuppressWarnings("unchecked")
     @Override
     public SimpleOption<Integer> newSlider(
             final String namespace,
@@ -95,8 +64,7 @@ public class Client119 extends Client1182 {
                     if (value.equals(min)) {
                         return GameOptions.getGenericValueText(optionText, Text.translatable(namespace + identifier + ".min"));
                     } else if (value.equals(max)) {
-                        return GameOptions.getGenericValueText(optionText,
-                                Text.translatable(namespace + identifier + ".max"));
+                        return GameOptions.getGenericValueText(optionText, Text.translatable(namespace + identifier + ".max"));
                     } else {
                         return GameOptions.getGenericValueText(optionText, value);
                     }
@@ -106,10 +74,5 @@ public class Client119 extends Client1182 {
                 Integer.parseInt(retrieveValueFunction.getValue(identifier)),
                 value -> storeValueFunction.setValue(identifier, Integer.toString(value))
         );
-    }
-
-    @Override
-    public boolean isWindowFocused() {
-        return getClientInstance().isWindowFocused();
     }
 }
