+++ Fastload-1.19.3-Extension/src/main/resources/fastload-1193-compat.mixins.json
@@ -5,7 +5,6 @@
   "plugin": "io.github.bumblesoftware.fastload.mixin.plugin.FLConditionalMixinConfig",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
-    "mc1193.local.HookClient"
   ],
   "client": [
     "mc1193.client.FrustumMixin",
