+++ Fastload-1.19.4-Extension/src/main/java/io/github/bumblesoftware/fastload/abstraction/client/mc1194/Client1194.java
@@ -1,6 +1,7 @@
-package io.github.bumblesoftware.fastload.abstraction.client;
+package io.github.bumblesoftware.fastload.abstraction.client.mc1194;
 
 
+import io.github.bumblesoftware.fastload.abstraction.client.mc1193.Client1193;
 import net.minecraft.client.font.TextRenderer;
 import net.minecraft.client.gui.DrawableHelper;
 import net.minecraft.client.gui.screen.Screen;
