+++ Fastload-1.19.4-Extension/src/main/resources/fastload-1194-compat.mixins.json
@@ -5,7 +5,6 @@
   "plugin": "io.github.bumblesoftware.fastload.mixin.plugin.FLConditionalMixinConfig",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
-    "mc1194.local.HookClient"
   ],
   "client": [
   ],
