--- src/main/java/io/github/bumblesoftware/fastload/api/external/abstraction/core/AbstractionApi.java
@@ -1,5 +1,0 @@
-package io.github.bumblesoftware.fastload.api.external.abstraction.core;
-
-import io.github.bumblesoftware.fastload.api.external.abstraction.tool.version.SupportedVersions;
-
-public interface AbstractionApi extends SupportedVersions { }
