+++ src/main/java/io/github/bumblesoftware/fastload/api/external/abstraction/core/config/RetrieveValueFunction.java
@@ -1,0 +1,5 @@
+package io.github.bumblesoftware.fastload.api.external.abstraction.core.config;
+
+public interface RetrieveValueFunction {
+    String getValue(String key);
+}
