+++ src/main/java/io/github/bumblesoftware/fastload/api/external/abstraction/core/config/StoreValueFunction.java
@@ -1,4 +1,4 @@
-package io.github.bumblesoftware.fastload.api.external.abstraction.tool.config;
+package io.github.bumblesoftware.fastload.api.external.abstraction.core.config;
 
 public interface StoreValueFunction {
     void setValue(String key, String value);
