+++ src/main/java/io/github/bumblesoftware/fastload/api/external/abstraction/core/handler/AbstractionEntrypoint.java
@@ -1,0 +1,8 @@
+package io.github.bumblesoftware.fastload.api.external.abstraction.core.handler;
+
+import io.github.bumblesoftware.fastload.api.external.events.AbstractEvent;
+import io.github.bumblesoftware.fastload.util.ObjectHolder;
+
+public interface AbstractionEntrypoint {
+    <A extends MethodAbstractionApi> void register(AbstractEvent<ObjectHolder<A>> clientAbstractionEvent);
+}
