+++ src/main/java/io/github/bumblesoftware/fastload/api/external/abstraction/core/handler/AbstractionFactory.java
@@ -1,7 +1,7 @@
-package io.github.bumblesoftware.fastload.api.external.abstraction.core;
+package io.github.bumblesoftware.fastload.api.external.abstraction.core.handler;
 
-import io.github.bumblesoftware.fastload.api.external.abstraction.core.AbstractionHandler.Environment;
-import io.github.bumblesoftware.fastload.api.external.abstraction.def.MinecraftAbstractionDirectoryManager;
+import io.github.bumblesoftware.fastload.api.external.abstraction.core.handler.AbstractionHandler.Environment;
+import io.github.bumblesoftware.fastload.api.external.abstraction.def.MinecraftAbstraction;
 import io.github.bumblesoftware.fastload.api.external.events.AbstractEvent;
 import io.github.bumblesoftware.fastload.util.ObjectHolder;
 
@@ -10,7 +10,7 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 
 public class AbstractionFactory {
-    public static <A extends AbstractionApi> AbstractionHandler<A> create(
+    public static <A extends MethodAbstractionApi> AbstractionHandler<A> create(
             final List<String> abstractionModIds,
             final Environment common,
             final Consumer<AbstractEvent<ObjectHolder<A>>> base,
@@ -19,11 +19,11 @@ public class AbstractionFactory {
         return new AbstractionHandler<>(abstractionModIds, common, base, abstractionInstanceGetter);
     }
 
-    public static <A extends AbstractionApi> AbstractionHandler<A> create(
+    public static <A extends MethodAbstractionApi> AbstractionHandler<A> create(
             final List<String> abstractionModIds,
             final Environment common,
             final Consumer<AbstractEvent<ObjectHolder<A>>> base
     ) {
-        return new AbstractionHandler<>(abstractionModIds, common, base, MinecraftAbstractionDirectoryManager::new);
+        return new AbstractionHandler<>(abstractionModIds, common, base, MinecraftAbstraction::new);
     }
 }
