+++ src/main/java/io/github/bumblesoftware/fastload/api/external/abstraction/core/handler/MethodAbstractionApi.java
@@ -1,0 +1,5 @@
+package io.github.bumblesoftware.fastload.api.external.abstraction.core.handler;
+
+import io.github.bumblesoftware.fastload.api.external.abstraction.core.versioning.SupportedVersions;
+
+public interface MethodAbstractionApi extends SupportedVersions {}
