+++ src/main/java/io/github/bumblesoftware/fastload/api/external/abstraction/core/versioning/VersionProvider.java
@@ -1,0 +1,5 @@
+package io.github.bumblesoftware.fastload.api.external.abstraction.core.versioning;
+
+public interface VersionProvider {
+    String getVersion();
+}
