+++ src/main/java/io/github/bumblesoftware/fastload/api/external/abstraction/core/versioning/VersionUtil.java
@@ -1,4 +1,4 @@
-package io.github.bumblesoftware.fastload.api.external.abstraction.tool.version;
+package io.github.bumblesoftware.fastload.api.external.abstraction.core.versioning;
 
 import io.github.bumblesoftware.fastload.init.Fastload;
 import net.fabricmc.loader.api.FabricLoader;
