--- src/main/java/io/github/bumblesoftware/fastload/api/external/abstraction/def/MinecraftVersion.java
@@ -1,5 +1,0 @@
-package io.github.bumblesoftware.fastload.api.external.abstraction.def;
-
-public interface MinecraftVersion {
-    String getVersion();
-}
