--- src/main/java/io/github/bumblesoftware/fastload/api/external/abstraction/tool/config/RetrieveValueFunction.java
@@ -1,5 +1,0 @@
-package io.github.bumblesoftware.fastload.api.external.abstraction.tool.config;
-
-public interface RetrieveValueFunction {
-    String getValue(String key);
-}
