+++ src/main/java/io/github/bumblesoftware/fastload/compat/modmenu/FastloadModMenuIntegration.java
@@ -5,10 +5,10 @@ import com.terraformersmc.modmenu.api.ModMenuApi;
 import io.github.bumblesoftware.fastload.api.internal.abstraction.AbstractClientCalls;
 import net.minecraft.client.gui.screen.Screen;
 
-import static io.github.bumblesoftware.fastload.init.FastloadClient.MINECRAFT_ABSTRACTION;
+import static io.github.bumblesoftware.fastload.init.FastloadClient.MINECRAFT_ABSTRACTION_HANDLER;
 
 public class FastloadModMenuIntegration implements ModMenuApi {
-    public static final AbstractClientCalls ABSTRACTED_CLIENT = MINECRAFT_ABSTRACTION.getAbstractedEntries();
+    public static final AbstractClientCalls ABSTRACTED_CLIENT = MINECRAFT_ABSTRACTION_HANDLER.directory.getAbstractedEntries();
     /**
      *  Registers our config screen to modmenu
      */
